remove irrelevant files
