+++ FaceUnity/android/.gradle/buildOutputCleanup/cache.properties
@@ -1,0 +1,2 @@
+#Fri Sep 29 19:41:59 CST 2023
+gradle.version=8.0
