--- FaceUnity/android/app/libs/PLACEHOLDER
@@ -1,1 +1,0 @@
-android-release.aar (https://download.agora.io/marketplace/release/Agora_Marketplace_FaceUnity_v8.6.0_Extension_for_Android_v4.1.1.zip)
