--- FaceUnity/android/app/src/main/assets/PLACEHOLDER
@@ -1,1 +1,0 @@
-/Resource (https://download.agora.io/marketplace/release/FaceUnity_v8.4.1_Resources.zip)
