--- FaceUnity/android/app/src/main/java/io/agora/rte/extension/faceunity/example/Config.java
@@ -1,6 +1,0 @@
-package io.agora.rte.extension.faceunity.example;
-
-public interface Config {
-    String mAppId = "<YOUR_APP_ID>";
-    String mToken = null;
-}
