--- FaceUnity/android/app/src/main/res/layout/activity_main.xml
@@ -1,115 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".MainActivity">
-
-    <SurfaceView
-        android:id="@+id/surfaceView"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <Button
-        android:id="@+id/button_enable"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginEnd="4dp"
-        android:layout_marginBottom="8dp"
-        android:text="@string/enable_extension"
-        android:textAllCaps="false"
-        app:layout_constraintBottom_toTopOf="@id/button_enable_ai"
-        app:layout_constraintEnd_toEndOf="parent" />
-
-    <Button
-        android:id="@+id/button_enable_ai"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginEnd="4dp"
-        android:layout_marginBottom="4dp"
-        android:text="enableAITracking"
-        android:textAllCaps="false"
-        app:layout_constraintBottom_toTopOf="@id/button_composers"
-        app:layout_constraintEnd_toEndOf="parent" />
-
-    <Button
-        android:id="@+id/button_stickers"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginEnd="4dp"
-        android:layout_marginBottom="8dp"
-        android:text="@string/set_sticker"
-        android:textAllCaps="false"
-        app:layout_constraintBottom_toTopOf="@+id/btn_lightmarkup"
-        app:layout_constraintEnd_toEndOf="parent" />
-
-    <Button
-        android:id="@+id/button_composers"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginEnd="4dp"
-        android:layout_marginBottom="4dp"
-        android:text="@string/set_composer"
-        android:textAllCaps="false"
-        app:layout_constraintBottom_toTopOf="@id/button_stickers"
-        app:layout_constraintEnd_toEndOf="parent" />
-
-    <Button
-        android:id="@+id/btn_lightmarkup"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginBottom="4dp"
-        android:text="@string/set_lightmarkup"
-        app:layout_constraintBottom_toTopOf="@+id/color_level"
-        app:layout_constraintEnd_toEndOf="parent" />
-
-    <androidx.appcompat.widget.AppCompatTextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        app:layout_constraintEnd_toStartOf="@+id/color_level"
-        app:layout_constraintTop_toTopOf="@+id/color_level"
-        app:layout_constraintBottom_toBottomOf="@+id/color_level"
-        android:text="Color Level"
-        />
-
-    <androidx.appcompat.widget.AppCompatSeekBar
-        android:id="@+id/color_level"
-        android:layout_width="100dp"
-        android:layout_height="50dp"
-        app:layout_constraintBottom_toTopOf="@+id/filter_level"
-        app:layout_constraintEnd_toEndOf="parent" />
-
-    <androidx.appcompat.widget.AppCompatTextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        app:layout_constraintEnd_toStartOf="@+id/filter_level"
-        app:layout_constraintTop_toTopOf="@+id/filter_level"
-        app:layout_constraintBottom_toBottomOf="@+id/filter_level"
-        android:text="Filter Level"
-        />
-
-    <androidx.appcompat.widget.AppCompatSeekBar
-        android:id="@+id/filter_level"
-        android:layout_width="100dp"
-        android:layout_height="50dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        />
-
-    <TextView
-        android:id="@+id/tv_ai_tracking"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        android:layout_marginEnd="16dp"
-        android:visibility="gone"
-        android:text="faces: 0\nhands: 0\npeople: 0"
-        android:layout_marginRight="16dp" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
