+++ FaceUnity/android/extension_demo/src/main/AndroidManifest.xml
