+++ FaceUnity/android/extension_demo/src/main/java/io/agora/rte/extension/faceunity/Config.java
@@ -1,0 +1,6 @@
+package io.agora.rte.extension.faceunity;
+
+public interface Config {
+    String mAppId = "1e46631b7a604c6ba31bdabb0259a58c" ; //TODO, Input your Agora AppId;
+    String mToken = null;
+}
